Moved card view-hillfort adapter code to separate file
