+++ app/src/main/java/org/wit/hillfortexplorer/activities/HillfortListActivity.kt
@@ -4,13 +4,11 @@ import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.view.*
 import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.recyclerview.widget.RecyclerView
 import kotlinx.android.synthetic.main.activity_hillfort_list.*
-import kotlinx.android.synthetic.main.card_hillfort.view.*
 import org.jetbrains.anko.startActivityForResult
 import org.wit.hillfortexplorer.R
 import org.wit.hillfortexplorer.main.MainApp
-import org.wit.hillfortexplorer.models.HillfortModel
+import org.wit.hillfortexplorer.models.HillfortAdapter
 
 class HillfortListActivity : AppCompatActivity() {
 
@@ -39,34 +37,5 @@ class HillfortListActivity : AppCompatActivity() {
             R.id.item_add -> startActivityForResult<HillfortActivity>(0)
         }
         return super.onOptionsItemSelected(item)
-    }
-}
-
-class HillfortAdapter constructor(private var hillforts: List<HillfortModel>):
-        RecyclerView.Adapter<HillfortAdapter.MainHolder>() {
-
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): MainHolder {
-        return MainHolder(
-            LayoutInflater.from(parent?.context).inflate(
-                R.layout.card_hillfort,
-                parent,
-                false
-            )
-        )
-    }
-
-    override fun onBindViewHolder(holder: MainHolder, position: Int) {
-        val hillfort = hillforts[holder.adapterPosition]
-        holder.bind(hillfort)
-    }
-
-    override fun getItemCount(): Int = hillforts.size
-
-    class MainHolder constructor(itemView: View): RecyclerView.ViewHolder(itemView) {
-
-        fun bind(hillfort: HillfortModel) {
-            itemView.hillfortTitle.text = hillfort.title
-            itemView.description.text = hillfort.description
-        }
     }
 }
