+++ app/src/main/java/org/wit/hillfortexplorer/models/HillfortAdapter.kt
@@ -1,0 +1,37 @@
+package org.wit.hillfortexplorer.models
+
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.recyclerview.widget.RecyclerView
+import kotlinx.android.synthetic.main.card_hillfort.view.*
+import org.wit.hillfortexplorer.R
+
+class HillfortAdapter constructor(private var hillforts: List<HillfortModel>):
+    RecyclerView.Adapter<HillfortAdapter.MainHolder>() {
+
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): MainHolder {
+        return MainHolder(
+            LayoutInflater.from(parent?.context).inflate(
+                R.layout.card_hillfort,
+                parent,
+                false
+            )
+        )
+    }
+
+    override fun onBindViewHolder(holder: MainHolder, position: Int) {
+        val hillfort = hillforts[holder.adapterPosition]
+        holder.bind(hillfort)
+    }
+
+    override fun getItemCount(): Int = hillforts.size
+
+    class MainHolder constructor(itemView: View): RecyclerView.ViewHolder(itemView) {
+
+        fun bind(hillfort: HillfortModel) {
+            itemView.hillfortTitle.text = hillfort.title
+            itemView.description.text = hillfort.description
+        }
+    }
+}
